Refactor calls2vcf to fix bubble decomp crash
Split calls2vcf code into three parts:
- decomp_bubble.{c,h}: bubble call entry + sam entry -> calls aligned to ref
- decomp_breakpoint.{c,h}: breakpoint call entry -> calls aligned to ref
- aligned_call.{c,h}: aligned call -> VCF entries

This allows us to use different strategies to realign breakpoint/bubble calls,
against the reference.

+++ src/tools/breakpoint_caller.c
@@ -10,7 +10,7 @@
 #include "json_hdr.h"
 
 typedef struct {
-  size_t first_runid, num_runs;
+  uint32_t first_runid, num_runs;
 } PathRefRun;
 
 typedef struct
